asynchandler: Pass Accept header to properly handle WFS3 request
By default, a QGIS server wfs3 request returns an html response if the
url does not contain an extension (html or json). However,
`py-qgis-server` returns a response as json by default. This is
because, QGIS server uses the `Accept` header to guess the correct
format if the extension is not set (See
`QgsServerOgcApiHandler::contentTypeFromRequest()`).

This issue is fixed by copying the `Accept` header from the initial
request to the request transmitted to QGIS server.